Fix fees on slot leader and common transactions

+++ GenesisBlock.cpp
@@ -18,11 +18,11 @@ GenesisBlock::GenesisBlock(){
 QList<std::pair<long,long>> const GenesisBlock::getPublicKeys(){
     return publicKey;
 }
-QList<std::pair<NodeClass*,int>> const GenesisBlock::getStakers(){
+QList<std::pair<NodeClass*,float>> const GenesisBlock::getStakers(){
     return stakers;
 }
 
-void GenesisBlock::addStaker(NodeClass* node, int stake){
+void GenesisBlock::addStaker(NodeClass* node, float stake){
     stakers.append(std::make_pair(node,stake));
 }
