+++ Transaction.cpp
@@ -34,6 +34,6 @@ void Transaction::setSignature(long long m_signature){
     signature = m_signature;
 }
 
-int Transaction::getFees(){
+float Transaction::getFees(){
     return fees;
 }
